HP-302 OIDC-client works with Tunnistamo
Config changes to make oidc client work with Tunnistamo. Tunnistamo does not support silentRenew like other endPoints so AutoSigIn config property must be false so renew is not attempted. Found also a bug when using callbacks, so callback routes are now separated in App.tsx or client would be initialized before callback and callback would not be called. An error will now be thrown is client init and callback are used together